+++ crates/rust-client/src/rpc/mod.rs
@@ -32,6 +32,8 @@ mod web_tonic_client;
 #[cfg(feature = "web-tonic")]
 pub use web_tonic_client::WebTonicRpcClient;
 
+use crate::sync::get_nullifier_prefix;
+
 // NOTE DETAILS
 // ================================================================================================
 
@@ -174,6 +176,27 @@ pub trait NodeRpcClient {
         block_num: u32,
         note_tags: &[NoteTag],
     ) -> Result<NoteSyncInfo, RpcError>;
+
+    /// Fetches the nullifiers corresponding to a list of prefixes using the
+    /// `/CheckNullifiersByPrefix` rpc endpoint
+    async fn check_nullifiers_by_prefix(
+        &mut self,
+        prefix: &[u16],
+    ) -> Result<Vec<(Nullifier, u32)>, RpcError>;
+
+    /// Fetches the commit height where the nullifier was consumed. If the nullifier is not found,
+    /// then `None` is returned.
+    ///
+    /// The default implementation of this method uses [NodeRpcClient::check_nullifiers_by_prefix].
+    async fn get_nullifier_commit_height(
+        &mut self,
+        nullifier: &Nullifier,
+    ) -> Result<Option<u32>, RpcError> {
+        let nullifiers =
+            self.check_nullifiers_by_prefix(&[get_nullifier_prefix(nullifier)]).await?;
+
+        Ok(nullifiers.iter().find(|(n, _)| n == nullifier).map(|(_, block_num)| *block_num))
+    }
 }
 
 // SYNC NOTE
@@ -292,6 +315,7 @@ impl CommittedNote {
 //
 #[derive(Debug)]
 pub enum NodeRpcClientEndpoint {
+    CheckNullifiersByPrefix,
     GetAccountDetails,
     GetBlockHeaderByNumber,
     SyncState,
@@ -302,6 +326,9 @@ pub enum NodeRpcClientEndpoint {
 impl fmt::Display for NodeRpcClientEndpoint {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            NodeRpcClientEndpoint::CheckNullifiersByPrefix => {
+                write!(f, "check_nullifiers_by_prefix")
+            },
             NodeRpcClientEndpoint::GetAccountDetails => write!(f, "get_account_details"),
             NodeRpcClientEndpoint::GetBlockHeaderByNumber => {
                 write!(f, "get_block_header_by_number")