+++ crates/rust-client/src/sync/mod.rs
@@ -8,7 +8,7 @@ use crypto::merkle::{InOrderIndex, MmrPeaks};
 use miden_objects::{
     accounts::{Account, AccountId, AccountStub},
     crypto::{self, rand::FeltRng},
-    notes::{Note, NoteId, NoteInclusionProof, NoteInputs, NoteRecipient, NoteTag},
+    notes::{Note, NoteId, NoteInclusionProof, NoteInputs, NoteRecipient, NoteTag, Nullifier},
     transaction::InputNote,
     BlockHeader, Digest,
 };
@@ -272,32 +272,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
             .map(|(acc_stub, _)| acc_stub)
             .collect();
 
-        let account_note_tags: Vec<NoteTag> = accounts
-            .iter()
-            .map(|acc| {
-                NoteTag::from_account_id(acc.id(), miden_objects::notes::NoteExecutionMode::Local)
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let stored_note_tags: Vec<NoteTag> = maybe_await!(self.store.get_note_tags())?;
-
-        let expected_notes = maybe_await!(self.store.get_input_notes(NoteFilter::Expected))?;
-
-        let uncommited_note_tags: Vec<NoteTag> = expected_notes
-            .iter()
-            .filter_map(|note| note.metadata().map(|metadata| metadata.tag()))
-            .collect();
-
-        let imported_tags: Vec<NoteTag> =
-            expected_notes.iter().filter_map(|note| note.imported_tag()).collect();
-
-        let note_tags: Vec<NoteTag> =
-            [account_note_tags, stored_note_tags, uncommited_note_tags, imported_tags]
-                .concat()
-                .into_iter()
-                .collect::<BTreeSet<NoteTag>>()
-                .into_iter()
-                .collect();
+        let note_tags: Vec<NoteTag> = maybe_await!(self.get_tracked_note_tags())?;
 
         // To receive information about added nullifiers, we reduce them to the higher 16 bits
         // Note that besides filtering by nullifier prefixes, the node also filters by block number
@@ -305,7 +280,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
         let nullifiers_tags: Vec<u16> =
             maybe_await!(self.store.get_unspent_input_note_nullifiers())?
                 .iter()
-                .map(|nullifier| (nullifier.inner()[3].as_int() >> FILTER_ID_SHIFT) as u16)
+                .map(get_nullifier_prefix)
                 .collect();
 
         // Send request
@@ -613,4 +588,8 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
         }
         Ok(())
     }
+}
+
+pub(crate) fn get_nullifier_prefix(nullifier: &Nullifier) -> u16 {
+    (nullifier.inner()[3].as_int() >> FILTER_ID_SHIFT) as u16
 }